cs: show documentation URL in description
