--- layout.go
@@ -1,8 +1,0 @@
-package egoscale
-
-// CommandInfo represents the meta data related to a Command
-type CommandInfo struct {
-	Name        string
-	Description string
-	RootOnly    bool
-}
